Added views.abc module to avoid repeated code
Updated type hints
