+++ vkbottle/dispatch/views/__init__.py
@@ -1,2 +1,1 @@
-from .abc import ABCView
-from .abc_dispense import ABCDispenseView
+from .abc import ABCDispenseView, ABCView
