+++ vkbottle/dispatch/views/abc/__init__.py
@@ -1,0 +1,4 @@
+from .dispense import ABCDispenseView
+from .message import ABCMessageView
+from .raw import ABCRawEventView
+from .view import ABCView
