+++ vkbottle/dispatch/views/abc/view.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Callable, List, Optional, Set, Type
+from typing import TYPE_CHECKING, List, Optional, Set, Type
 
 from vkbottle.api.abc import ABCAPI
 from vkbottle.dispatch.dispenser.abc import ABCStateDispenser
@@ -7,7 +7,6 @@ from vkbottle.dispatch.handlers import ABCHandler
 from vkbottle.dispatch.middlewares import BaseMiddleware
 from vkbottle.dispatch.return_manager import BaseReturnManager
 from vkbottle.modules import logger
-from vkbottle.tools.dev_tools.mini_types.user.message import MessageMin
 
 if TYPE_CHECKING:
     from vkbottle_types.events import Event
@@ -20,8 +19,6 @@ class ABCView(ABC):
     middlewares: Set[Type["BaseMiddleware"]]
     middleware_instances: List["BaseMiddleware"]
     handler_return_manager: BaseReturnManager
-    state_source_key: str
-    default_text_approximators: List[Callable[[MessageMin], str]]
 
     @abstractmethod
     def __init__(self):
@@ -29,8 +26,6 @@ class ABCView(ABC):
         self.middlewares = set()
         self.middleware_instances = []
         self.handler_return_manager = None
-        self.state_source_key = DEFAULT_STATE_KEY
-        self.default_text_approximators = []
 
     @abstractmethod
     async def process_event(self, event: "Event") -> bool:
