+++ vkbottle/dispatch/views/bot/message.py
@@ -1,70 +1,29 @@
-from abc import ABC
 from typing import Optional
 
 from vkbottle_types.events import GroupEventType
 
-from vkbottle.api.abc import ABCAPI
-from vkbottle.dispatch.dispenser.abc import ABCStateDispenser
 from vkbottle.dispatch.return_manager.bot import BotMessageReturnHandler
-from vkbottle.modules import logger
+from vkbottle.dispatch.views.abc import ABCMessageView
 from vkbottle.tools.dev_tools.mini_types.bot import message_min
 
-from ..abc_dispense import ABCDispenseView
 
-
-class ABCMessageView(ABCDispenseView, ABC):
+class ABCBotMessageView(ABCMessageView):
     def __init__(self):
         super().__init__()
         self.handler_return_manager = BotMessageReturnHandler()
 
+    @staticmethod
+    def get_logger_event_value(event):
+        return event.get("event_id")
+
+    @staticmethod
+    async def get_message(event, ctx_api):
+        return message_min(event, ctx_api)
+
     async def process_event(self, event: dict) -> bool:
         return GroupEventType(event["type"]) == GroupEventType.MESSAGE_NEW
 
-    async def handle_event(
-        self, event: dict, ctx_api: "ABCAPI", state_dispenser: "ABCStateDispenser"
-    ) -> None:
-        logger.debug("Handling event ({}) with message view".format(event.get("event_id")))
-        context_variables: dict = {}
-        message = message_min(event, ctx_api)  # TODO
-        message.state_peer = await state_dispenser.cast(self.get_state_key(event))
 
-        for text_ax in self.default_text_approximators:
-            message.text = text_ax(message)
-
-        error = await self.pre_middleware(message, context_variables)
-        if error:
-            logger.info("Handling stopped, pre_middleware returned error")
-            return
-
-        handle_responses = []
-        handlers = []
-
-        for handler in self.handlers:
-            result = await handler.filter(message)
-            logger.debug("Handler {} returned {}".format(handler, result))
-
-            if result is False:
-                continue
-
-            elif isinstance(result, dict):
-                context_variables.update(result)
-
-            handler_response = await handler.handle(message, **context_variables)
-            handle_responses.append(handler_response)
-            handlers.append(handler)
-
-            return_handler = self.handler_return_manager.get_handler(handler_response)
-            if return_handler is not None:
-                await return_handler(
-                    self.handler_return_manager, handler_response, message, context_variables
-                )
-
-            if handler.blocking:
-                break
-
-        await self.post_middleware(handle_responses, handlers)
-
-
-class MessageView(ABCMessageView):
+class MessageView(ABCBotMessageView):
     def get_state_key(self, event: dict) -> Optional[int]:
         return event["object"]["message"].get(self.state_source_key)
